+++ testcase/mydtype.py
@@ -4,12 +4,12 @@ import numpy as np
 
 test_conf=np.dtype([('jpi'   ,int)  ,('jpj',int)     ,('jpk',int),\
                     ('nprocx',int)  ,('nprocy',int)  ,\
-                    ('lon0'  ,np.float),('lat0',np.float)  ,\
-                    ('dx'    ,np.float),('dy'    ,np.float),\
+                    ('lon0'  ,float),('lat0',float)  ,\
+                    ('dx'    ,float),('dy'    ,float),\
                     ('Start' ,'S100')  ,('End'   ,'S100')  ,\
                     ('Dir','S100'),('Code','S100')])
 
-ext_data =np.dtype([('date','S17'),('kext',np.float)])
+ext_data =np.dtype([('date','S17'),('kext',float)])
 
 
 def file2stringlist(filename):
